refactor(cli): integrate core encoding utilities; add reset function for testing
Includes function comments and types for clarity and a cache reset utility for test support.
