+++ packages/core/src/index.ts
@@ -35,6 +35,7 @@ export * from './utils/editor.js';
 export * from './utils/quotaErrorDetection.js';
 export * from './utils/fileUtils.js';
 export * from './utils/retry.js';
+export * from './utils/systemEncoding.js';
 
 // Export services
 export * from './services/fileDiscoveryService.js';
